+++ h2gis-functions/src/main/java/org/h2gis/functions/spatial/properties/ST_EstimatedExtent.java
@@ -7,7 +7,6 @@ package org.h2gis.functions.spatial.properties;
 
 import org.h2gis.api.AbstractFunction;
 import org.h2gis.api.ScalarFunction;
-import org.h2gis.utilities.SFSUtilities;
 import org.h2gis.utilities.TableLocation;
 import org.locationtech.jts.geom.Geometry;
