+++ h2gis-functions/src/main/java/org/h2gis/functions/spatial/topography/ST_TriangleContouring.java
@@ -27,7 +27,6 @@ import org.h2.value.ValueArray;
 import org.h2.value.ValueVarchar;
 import org.h2gis.api.DeterministicScalarFunction;
 import org.h2gis.utilities.JDBCUtilities;
-import org.h2gis.utilities.SFSUtilities;
 import org.h2gis.utilities.TableLocation;
 import org.h2gis.utilities.TableUtilities;
 import org.h2gis.utilities.jts_utils.Contouring;
