+++ h2gis-functions/src/test/java/org/h2gis/functions/BasicTest.java
@@ -25,7 +25,6 @@ import org.h2.value.DataType;
 import org.h2.value.Value;
 import org.h2gis.functions.factory.H2GISDBFactory;
 import org.h2gis.functions.factory.H2GISFunctions;
-import org.h2gis.utilities.SFSUtilities;
 import org.h2gis.utilities.TableLocation;
 import org.h2gis.utilities.trigger.UpdateTrigger;
 import org.junit.jupiter.api.*;
@@ -36,7 +35,6 @@ import org.locationtech.jts.io.ParseException;
 import org.locationtech.jts.io.WKTReader;
 
 import java.sql.*;
-import java.util.List;
 import org.h2gis.utilities.GeometryTableUtilities;
 import org.h2gis.utilities.Tuple;
