+++ h2gis-utilities/src/test/java/org/h2gis/utilities/GeographyUtilsTest.java
@@ -19,7 +19,6 @@
  */
 package org.h2gis.utilities;
 
-import org.h2gis.utilities.jts_utils.GeographyUtils;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -37,14 +36,14 @@ public class GeographyUtilsTest {
     @Test
     public void expandEnvelopeByMeters1() throws Exception {
         Envelope env = new Envelope(0.0, 2.0, 1.0, 2.0);
-        assertEquals(env, GeographyUtils.expandEnvelopeByMeters(env, 0));
+        assertEquals(env, GeographyUtilities.expandEnvelopeByMeters(env, 0));
     }
 
     @Test
     public void expandEnvelopeByMeters2() throws Exception {
         Envelope env = new Envelope(0.0, 2.0, 1.0, 2.0);
-        Envelope expandedEnv = GeographyUtils.expandEnvelopeByMeters(env, 1000);
-        assertEquals(1000, GeographyUtils.getHaversineDistanceInMeters(new Coordinate(env.getMinX(), env.getMinY()),
+        Envelope expandedEnv = GeographyUtilities.expandEnvelopeByMeters(env, 1000);
+        assertEquals(1000, GeographyUtilities.getHaversineDistanceInMeters(new Coordinate(env.getMinX(), env.getMinY()),
                 new Coordinate(expandedEnv.getMinX(), env.getMinY())), 1, "Expanded to 1000");
     }
 
@@ -52,7 +51,7 @@ public class GeographyUtilsTest {
     public void expandEnvelopeByMeters3() throws Exception {
         Envelope env = new Envelope(22220.0, 2.0, 1.0, 2.0);
         assertThrows(IllegalArgumentException.class, () -> {
-            assertEquals(env, GeographyUtils.expandEnvelopeByMeters(env, 1000));
+            assertEquals(env, GeographyUtilities.expandEnvelopeByMeters(env, 1000));
         });
     }
 
@@ -60,15 +59,15 @@ public class GeographyUtilsTest {
     public void expandEnvelopeByMeters4() throws Exception {
         Envelope env = new Envelope(0.0, 2.0, 1.0, 2.0);
         assertThrows(IllegalArgumentException.class, () -> {
-            assertEquals(env, GeographyUtils.expandEnvelopeByMeters(env, -1));
+            assertEquals(env, GeographyUtilities.expandEnvelopeByMeters(env, -1));
         });
     }
 
     @Test
     public void expandEnvelopeByMeters5() throws Exception {
         Envelope env = new Envelope(0.0, .0, 0.0, 0.0);
-        Envelope expandedEnv = GeographyUtils.expandEnvelopeByMeters(env, 200);
-        assertEquals(200, GeographyUtils.getHaversineDistanceInMeters(new Coordinate(env.getMinX(), env.getMinY()),
+        Envelope expandedEnv = GeographyUtilities.expandEnvelopeByMeters(env, 200);
+        assertEquals(200, GeographyUtilities.getHaversineDistanceInMeters(new Coordinate(env.getMinX(), env.getMinY()),
                 new Coordinate(expandedEnv.getMinX(), env.getMinY())), 1, "Expanded to 200");
     }
     
@@ -77,7 +76,7 @@ public class GeographyUtilsTest {
     public void haversineDistanceInMeters1() throws Exception {
         Coordinate coorda = new Coordinate(-77.037852, 38.898556);
         Coordinate coordb = new Coordinate(-77.043934, 38.897147);
-        assertEquals(549, GeographyUtils.getHaversineDistanceInMeters(coorda, coordb), 1, "HaversineDistance");
+        assertEquals(549, GeographyUtilities.getHaversineDistanceInMeters(coorda, coordb), 1, "HaversineDistance");
     }
 
     /**
@@ -92,22 +91,22 @@ public class GeographyUtilsTest {
     
     @Test
     public void computeLatitudeDistance1() throws Exception {
-        assertEquals(0, GeographyUtils.computeLatitudeDistance(0));
+        assertEquals(0, GeographyUtilities.computeLatitudeDistance(0));
     }
     
     @Test
     public void computeLatitudeDistance2() throws Exception {
-        assertTrue((1- GeographyUtils.computeLatitudeDistance(111.045))>0.01);
+        assertTrue((1- GeographyUtilities.computeLatitudeDistance(111.045))>0.01);
     }
     
     @Test
     public void computeLongitudeDistance1() throws Exception {
-        assertEquals(0, GeographyUtils.computeLongitudeDistance(0, 0));
+        assertEquals(0, GeographyUtilities.computeLongitudeDistance(0, 0));
     }
     
     @Test
     public void computeLongitudeDistance2() throws Exception {
-        assertTrue((1- GeographyUtils.computeLongitudeDistance(87.87018,3 ))>0.01);
+        assertTrue((1- GeographyUtilities.computeLongitudeDistance(87.87018,3 ))>0.01);
     }
 
 }
