+++ h2gis-utilities/src/test/java/org/h2gis/utilities/GeometryMetadataTest.java
@@ -28,7 +28,6 @@ import org.locationtech.jts.geom.CoordinateXY;
 import org.locationtech.jts.geom.CoordinateXYM;
 import org.locationtech.jts.geom.CoordinateXYZM;
 import org.locationtech.jts.geom.GeometryFactory;
-import org.locationtech.jts.io.ParseException;
 
 /**
  *
